# Optimize flatMapConcat for single element source, #25241