* This first stab is only looking for `SingleSource` and note
  that `Source.single` is not returning that. Source.single has
  wrap with fromGraph and then the original `SingleSource` is
  lost.

Refs #25241

Benchmark results speak for themselves, 287k vs 6641k elements/s:
```
jmh:run -prof gc -i 5 -wi 5 -f2 -t1 akka.stream.FlatMapConcatBenchmark

[info] FlatMapConcatBenchmark.single                                        thrpt   10    287130.513 ±    4748.851   ops/s
[info] FlatMapConcatBenchmark.single:·gc.alloc.rate                         thrpt   10      1327.483 ±    1058.667  MB/sec
[info] FlatMapConcatBenchmark.single:·gc.alloc.rate.norm                    thrpt   10      5095.151 ±    4060.888    B/op

[info] FlatMapConcatBenchmark.optimizedSingle                               thrpt   10   6641485.752 ±   42696.132   ops/s
[info] FlatMapConcatBenchmark.optimizedSingle:·gc.alloc.rate                thrpt   10       502.781 ±     400.356  MB/sec
[info] FlatMapConcatBenchmark.optimizedSingle:·gc.alloc.rate.norm           thrpt   10        83.397 ±      66.405    B/op

[info] Benchmark                                                             Mode  Cnt         Score         Error   Units
[info] FlatMapConcatBenchmark.mapBaseline                                   thrpt   10  14275746.640 ± 1653684.721   ops/s
[info] FlatMapConcatBenchmark.mapBaseline:·gc.alloc.rate                    thrpt   10         1.847 ±       1.005  MB/sec
[info] FlatMapConcatBenchmark.mapBaseline:·gc.alloc.rate.norm               thrpt   10         0.143 ±       0.076    B/op
```