@patriknw 2018-06-20
I [found a way](https://github.com/akka/akka/pull/25242/commits/cf9f96323b66f88142b389dbb326ce7c08b0453f) to grab the original `SingleSource` via the TraversalBuilder.

`flatMapConcat(Source.single)` improves from 295k ops/s to 3697k ops/s.

We could get an additional boost to 6743k ops/s by using the `SourceSingle` without the wrapping in `fromGraph`. That might not be needed. I think it's more than enough for the websocket case anyway. If we want to do that I prefer to add it as public API.

```
  def singleElementGraph[T](element: T): Graph[SourceShape[T], NotUsed] =
    new GraphStages.SingleSource(element)
```

I also verified that the extra work doesn't introduce a regression for non single, see `oneElementList` in results below.

BEFORE

```
[info] Benchmark                                                                  Mode  Cnt         Score         Error   Units
[info] FlatMapConcatBenchmark.internalSingleSource                               thrpt   10    304469.550 ±    3026.728   ops/s
[info] FlatMapConcatBenchmark.internalSingleSource:·gc.alloc.rate                thrpt   10      1394.921 ±    1111.594  MB/sec
[info] FlatMapConcatBenchmark.internalSingleSource:·gc.alloc.rate.norm           thrpt   10      5043.950 ±    4019.065    B/op

[info] FlatMapConcatBenchmark.mapBaseline                                        thrpt   10  14286020.955 ± 1613689.270   ops/s
[info] FlatMapConcatBenchmark.mapBaseline:·gc.alloc.rate                         thrpt   10         1.875 ±       1.050  MB/sec
[info] FlatMapConcatBenchmark.mapBaseline:·gc.alloc.rate.norm                    thrpt   10         0.142 ±       0.076    B/op

[info] FlatMapConcatBenchmark.oneElementList                                     thrpt   10    222884.964 ±   14273.282   ops/s
[info] FlatMapConcatBenchmark.oneElementList:·gc.alloc.rate                      thrpt   10      1246.652 ±     994.710  MB/sec
[info] FlatMapConcatBenchmark.oneElementList:·gc.alloc.rate.norm                 thrpt   10      6093.632 ±    4855.525    B/op

[info] FlatMapConcatBenchmark.sourceDotSingle                                    thrpt   10    295337.449 ±   14037.933   ops/s
[info] FlatMapConcatBenchmark.sourceDotSingle:·gc.alloc.rate                     thrpt   10      1361.842 ±    1088.199  MB/sec
[info] FlatMapConcatBenchmark.sourceDotSingle:·gc.alloc.rate.norm                thrpt   10      5063.151 ±    4034.767    B/op
```

AFTER

```
[info] Benchmark                                                                  Mode  Cnt         Score        Error   Units
[info] FlatMapConcatBenchmark.internalSingleSource                               thrpt   10   6743400.176 ±  30127.278   ops/s
[info] FlatMapConcatBenchmark.internalSingleSource:·gc.alloc.rate                thrpt   10       510.081 ±    406.160  MB/sec
[info] FlatMapConcatBenchmark.internalSingleSource:·gc.alloc.rate.norm           thrpt   10        83.397 ±     66.405    B/op

[info] FlatMapConcatBenchmark.mapBaseline                                        thrpt   10  14618585.353 ±  26890.408   ops/s
[info] FlatMapConcatBenchmark.mapBaseline:·gc.alloc.rate                         thrpt   10         1.904 ±      1.010  MB/sec
[info] FlatMapConcatBenchmark.mapBaseline:·gc.alloc.rate.norm                    thrpt   10         0.143 ±      0.076    B/op

[info] FlatMapConcatBenchmark.oneElementList                                     thrpt   10    228616.358 ±  22694.113   ops/s
[info] FlatMapConcatBenchmark.oneElementList:·gc.alloc.rate                      thrpt   10      1260.890 ±   1016.359  MB/sec
[info] FlatMapConcatBenchmark.oneElementList:·gc.alloc.rate.norm                 thrpt   10      6084.032 ±   4847.906    B/op

[info] FlatMapConcatBenchmark.sourceDotSingle                                    thrpt   10   3697623.899 ± 124067.029   ops/s
[info] FlatMapConcatBenchmark.sourceDotSingle:·gc.alloc.rate                     thrpt   10      1010.126 ±    804.724  MB/sec
[info] FlatMapConcatBenchmark.sourceDotSingle:·gc.alloc.rate.norm                thrpt   10       301.000 ±    240.064    B/op
```