@patriknw 2018-06-20
Pushed another change to take care of the case when there is no demand https://github.com/akka/akka/pull/25242/commits/b4ea15e0c005c4571e1f0ac90d43e0c8632636b8